+++ src/main/java/net/additional_jewelry/items/Items.java
@@ -3,6 +3,7 @@ package net.additional_jewelry.items;
 import net.additional_jewelry.AdditionalJewelry;
 import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
 import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
+import net.fabricmc.loader.api.FabricLoader;
 import net.jewelry.api.AttributeResolver;
 import net.jewelry.api.JewelryItem;
 import net.jewelry.config.ItemConfig;
@@ -17,68 +18,36 @@ import net.spell_power.api.SpellPowerMechanics;
 
 import java.util.*;
 
-public class Items {
-    public record Entry(Identifier id, JewelryItem item, ItemConfig.Item config) {
-    }
+import static net.jewelry.items.JewelryItems.GENERIC_ATTACK_DAMAGE;
+import static net.jewelry.items.JewelryItems.GENERIC_MAX_HEALTH;
 
+public class Items {
+    public record Entry(Identifier id, JewelryItem item, ItemConfig.Item config, boolean fireproof) {  }
     public static final ArrayList<Entry> all = new ArrayList<>();
     public static final Map<String, Item> entryMap = new HashMap<>();
 
     public static Entry add(Identifier id, ItemConfig.Item config) {
-        return add(id, Rarity.COMMON, config, null);
+        return add(id, Rarity.COMMON, config, null, false);
     }
 
     public static Entry add(Identifier id, Rarity rarity, ItemConfig.Item config) {
-        return add(id, rarity, config, null);
+        return add(id, rarity, config, null, false);
+    }
+
+    public static Entry add(Identifier id, Rarity rarity, ItemConfig.Item config, boolean fireproof) {
+        return add(id, rarity, config, null, fireproof);
     }
 
     public static Entry add(Identifier id, Rarity rarity, boolean addLore, ItemConfig.Item config) {
-        return add(id, rarity, config, addLore ? ("item." + id.getNamespace() + "." + id.getPath() + ".lore") : null);
+        return add(id, rarity, config, addLore ? ("item." + id.getNamespace() + "." + id.getPath() + ".lore") : null, false);
     }
 
-
-    public static Entry add(Identifier id, Rarity rarity, ItemConfig.Item config, String lore) {
-        var entry = new Entry(id, new JewelryItem(new FabricItemSettings().rarity(rarity), lore), config);
-        all.add(entry);
-        entryMap.put(id.toString(), entry.item());
-        return entry;
-    }
-
-    ////RAGE
-    public static Entry addRage(Identifier id, Rarity rarity, ItemConfig.Item config, float value) {
-        return addRage(id, rarity, config, null, value);
-    }
-    public static Entry addRage(Identifier id, Rarity rarity, boolean addLore,ItemConfig.Item config, float value) {
-        return addRage(id, rarity, config, addLore ? ("item." + id.getNamespace() + "." + id.getPath() + ".lore") : null, value);
-    }
-    public static Entry addRage(Identifier id, Rarity rarity, ItemConfig.Item config, String lore, float value) {
-        var entry = new Entry(id, new RageJewelryItem(new FabricItemSettings().rarity(rarity), lore, value), config);
-        all.add(entry);
-        entryMap.put(id.toString(), entry.item());
-        return entry;
-    }
-    ///DAMAGEREFLECT
-    public static Entry addReflect(Identifier id, ItemConfig.Item config, float value) {
-        return addReflect(id, Rarity.COMMON, config, value);
-    }
-    public static Entry addReflect(Identifier id, Rarity rarity, ItemConfig.Item config, float value) {
-        return addReflect(id, rarity, config, null, value);
-    }
-    public static Entry addReflect(Identifier id, Rarity rarity, boolean addLore, ItemConfig.Item config,float value) {
-        return add(id, rarity, config, addLore ? ("item." + id.getNamespace() + "." + id.getPath() + ".lore") : null);
-    }
-    public static Entry addReflect(Identifier id, Rarity rarity, ItemConfig.Item config, String lore, float value) {
-        var entry = new Entry(id, new ReflectJewelryItem(new FabricItemSettings().rarity(rarity), lore, value), config);
-        all.add(entry);
-        entryMap.put(id.toString(), entry.item());
-        return entry;
-    }
-    //LIFESTEAL
-    public static Entry addLifeSteal(Identifier id, Rarity rarity, boolean addLore, ItemConfig.Item config,float value) {
-        return addLifeSteal(id, rarity, config, addLore ? ("item." + id.getNamespace() + "." + id.getPath() + ".lore") : null, value);
-    }
-    public static Entry addLifeSteal(Identifier id, Rarity rarity, ItemConfig.Item config, String lore, float value) {
-        var entry = new Entry(id, new LifeStealJewelryItem(new FabricItemSettings().rarity(rarity), lore, value), config);
+    public static Entry add(Identifier id, Rarity rarity, ItemConfig.Item config, String lore, boolean fireproof) {
+        var settings = new FabricItemSettings().rarity(rarity);
+        if (fireproof) {
+            settings = settings.fireproof();
+        }
+        var entry = new Entry(id, new JewelryItem(settings, lore), config, fireproof);
         all.add(entry);
         entryMap.put(id.toString(), entry.item());
         return entry;
@@ -95,14 +64,20 @@ public class Items {
 
 
     //CUSTOM_JEWELRY
-    public static Entry driptstone_necklace = addReflect(new Identifier(AdditionalJewelry.MOD_ID, "dripstone_necklace"), new ItemConfig.Item(
+    public static Entry driptstone_necklace = add(new Identifier(AdditionalJewelry.MOD_ID, "dripstone_necklace"), new ItemConfig.Item(
             List.of(
-            )
-    ),tier_1_multiplier);
-    public static Entry cactea_ring = addReflect(new Identifier(AdditionalJewelry.MOD_ID, "cactea_ring"), new ItemConfig.Item(
+                    new ItemConfig.AttributeModifier(GENERIC_MAX_HEALTH, 2, EntityAttributeModifier.Operation.ADDITION),
+                    new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                            Identifier.tryParse("more_rpg_classes:damage_reflect_modifier")),
+                            0.1F, EntityAttributeModifier.Operation.MULTIPLY_BASE))
+    ));
+    public static Entry cactea_ring = add(new Identifier(AdditionalJewelry.MOD_ID, "cactea_ring"), new ItemConfig.Item(
             List.of(
-            )
-    ),tier_1_multiplier);
+                    new ItemConfig.AttributeModifier(GENERIC_MAX_HEALTH, 2, EntityAttributeModifier.Operation.ADDITION),
+                    new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                            Identifier.tryParse("more_rpg_classes:damage_reflect_modifier")),
+                            0.1F, EntityAttributeModifier.Operation.MULTIPLY_BASE))
+    ));
 
     //UNCOMMON
     public static Entry ocean_ring = add(new Identifier(AdditionalJewelry.MOD_ID, "ocean_ring"), Rarity.UNCOMMON, new ItemConfig.Item(
@@ -120,10 +95,13 @@ public class Items {
                     new ItemConfig.AttributeModifier(MoreSpellSchools.EARTH.id, tier_1_bonus)
             )
     ));
-    public static Entry rage_ring = addRage(new Identifier(AdditionalJewelry.MOD_ID, "rage_ring"), Rarity.UNCOMMON, new ItemConfig.Item(
+    public static Entry rage_ring = add(new Identifier(AdditionalJewelry.MOD_ID, "rage_ring"), Rarity.UNCOMMON, new ItemConfig.Item(
             List.of(
+                    new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                            Identifier.tryParse("more_rpg_classes:rage_modifier")),
+                            0.1F, EntityAttributeModifier.Operation.MULTIPLY_BASE)
             )
-    ),tier_1_multiplier);
+    ));
 
     public static Entry ocean_necklace = add(new Identifier(AdditionalJewelry.MOD_ID, "ocean_necklace"), Rarity.UNCOMMON, new ItemConfig.Item(
             List.of(
@@ -140,10 +118,13 @@ public class Items {
                     new ItemConfig.AttributeModifier(MoreSpellSchools.EARTH.id, tier_1_bonus)
             )
     ));
-    public static Entry rage_necklace = addRage(new Identifier(AdditionalJewelry.MOD_ID, "rage_necklace"), Rarity.UNCOMMON, new ItemConfig.Item(
+    public static Entry rage_necklace = add(new Identifier(AdditionalJewelry.MOD_ID, "rage_necklace"), Rarity.UNCOMMON, new ItemConfig.Item(
             List.of(
+                    new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                            Identifier.tryParse("more_rpg_classes:rage_modifier")),
+                            0.1F, EntityAttributeModifier.Operation.MULTIPLY_BASE)
             )
-    ),tier_1_multiplier);
+    ));
 
     //NETHERITE_VERSIONS
     public static Entry netherite_ocean_ring = add(new Identifier(AdditionalJewelry.MOD_ID, "netherite_ocean_ring"), Rarity.UNCOMMON, new ItemConfig.Item(
@@ -161,10 +142,14 @@ public class Items {
                     new ItemConfig.AttributeModifier(MoreSpellSchools.EARTH.id, tier_2_bonus)
             )
     ));
-    public static Entry netherite_rage_ring = addRage(new Identifier(AdditionalJewelry.MOD_ID, "netherite_rage_ring"), Rarity.UNCOMMON, new ItemConfig.Item(
+    public static Entry netherite_rage_ring = add(new Identifier(AdditionalJewelry.MOD_ID, "netherite_rage_ring"), Rarity.UNCOMMON, new ItemConfig.Item(
             List.of(
+                    new ItemConfig.AttributeModifier("generic.attack_damage", 0.05F, EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                    new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                            Identifier.tryParse("more_rpg_classes:rage_modifier")),
+                            0.15F, EntityAttributeModifier.Operation.MULTIPLY_BASE)
             )
-    ),tier_2_multiplier);
+    ));
 
     public static Entry netherite_ocean_necklace = add(new Identifier(AdditionalJewelry.MOD_ID, "netherite_ocean_necklace"), Rarity.UNCOMMON, new ItemConfig.Item(
             List.of(
@@ -181,10 +166,14 @@ public class Items {
                     new ItemConfig.AttributeModifier(MoreSpellSchools.EARTH.id, tier_2_bonus)
             )
     ));
-    public static Entry netherite_rage_necklace = addRage(new Identifier(AdditionalJewelry.MOD_ID, "netherite_rage_necklace"), Rarity.UNCOMMON, new ItemConfig.Item(
+    public static Entry netherite_rage_necklace = add(new Identifier(AdditionalJewelry.MOD_ID, "netherite_rage_necklace"), Rarity.UNCOMMON, new ItemConfig.Item(
             List.of(
+                    new ItemConfig.AttributeModifier("generic.attack_damage", 0.05F, EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                    new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                            Identifier.tryParse("more_rpg_classes:rage_modifier")),
+                            0.15F, EntityAttributeModifier.Operation.MULTIPLY_BASE)
             )
-    ),tier_2_multiplier);
+    ));
 
 
     //UNIQUE
@@ -202,7 +191,6 @@ public class Items {
                     new ItemConfig.AttributeModifier(SpellPowerMechanics.HASTE.id, tier_3_secondary_multiplier, EntityAttributeModifier.Operation.MULTIPLY_BASE)
             )
     ));
-
     public static Entry unique_sky_ring = add(new Identifier(AdditionalJewelry.MOD_ID, "unique_sky_ring"), Rarity.RARE, true, new ItemConfig.Item(
             List.of(
                     new ItemConfig.AttributeModifier(MoreSpellSchools.AIR.id, tier_3_spell_multiplier, EntityAttributeModifier.Operation.MULTIPLY_BASE),
@@ -217,7 +205,6 @@ public class Items {
                     new ItemConfig.AttributeModifier(SpellPowerMechanics.HASTE.id, tier_3_secondary_multiplier, EntityAttributeModifier.Operation.MULTIPLY_BASE)
             )
     ));
-
     public static Entry unique_earth_ring = add(new Identifier(AdditionalJewelry.MOD_ID, "unique_earth_ring"), Rarity.RARE, true, new ItemConfig.Item(
             List.of(
                     new ItemConfig.AttributeModifier(MoreSpellSchools.EARTH.id, tier_3_spell_multiplier, EntityAttributeModifier.Operation.MULTIPLY_BASE),
@@ -232,28 +219,136 @@ public class Items {
                     new ItemConfig.AttributeModifier(SpellPowerMechanics.CRITICAL_DAMAGE.id, tier_3_secondary_multiplier, EntityAttributeModifier.Operation.MULTIPLY_BASE)
             )
     ));
+    public static Entry unique_rage_ring = add(new Identifier(AdditionalJewelry.MOD_ID, "unique_rage_ring"), Rarity.RARE, true, new ItemConfig.Item(
+            List.of(
+                    new ItemConfig.AttributeModifier("generic.attack_damage", 0.1F, EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                    new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                            Identifier.tryParse("more_rpg_classes:rage_modifier")),
+                            0.2F, EntityAttributeModifier.Operation.MULTIPLY_BASE)
+            )
+    ));
+    public static Entry unique_rage_necklace = add(new Identifier(AdditionalJewelry.MOD_ID, "unique_rage_necklace"), Rarity.RARE, true ,new ItemConfig.Item(
+            List.of(
+                    new ItemConfig.AttributeModifier("generic.attack_damage", 0.1F, EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                    new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                            Identifier.tryParse("more_rpg_classes:rage_modifier")),
+                            0.2F, EntityAttributeModifier.Operation.MULTIPLY_BASE)
+            )
+    ));
+    public static Entry vampire_ring = add(new Identifier(AdditionalJewelry.MOD_ID, "vampire_ring"), Rarity.RARE, true, new ItemConfig.Item(
+            List.of(
+                    new ItemConfig.AttributeModifier(GENERIC_ATTACK_DAMAGE, 0.08F, EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                    new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                            Identifier.tryParse("more_rpg_classes:lifesteal_modifier")),
+                            lifesteal_multiplier, EntityAttributeModifier.Operation.MULTIPLY_BASE)
+            )
+    ));
+    public static Entry vampire_necklace = add(new Identifier(AdditionalJewelry.MOD_ID, "vampire_necklace"), Rarity.RARE, true ,new ItemConfig.Item(
+            List.of(
+                    new ItemConfig.AttributeModifier(GENERIC_ATTACK_DAMAGE, 0.08F, EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                    new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                            Identifier.tryParse("more_rpg_classes:lifesteal_modifier")),
+                            lifesteal_multiplier, EntityAttributeModifier.Operation.MULTIPLY_BASE)
+            )
+    ));
 
-
-    public static Entry unique_rage_ring = addRage(new Identifier(AdditionalJewelry.MOD_ID, "unique_rage_ring"), Rarity.RARE, true, new ItemConfig.Item(
-            List.of(
-                    new ItemConfig.AttributeModifier("generic.attack_damage", tier_3_physical_multiplier/ 2, EntityAttributeModifier.Operation.MULTIPLY_BASE)
-            )
-    ),tier_3_spell_multiplier);
-    public static Entry unique_rage_necklace = addRage(new Identifier(AdditionalJewelry.MOD_ID, "unique_rage_necklace"), Rarity.RARE, true ,new ItemConfig.Item(
-            List.of(
-                    new ItemConfig.AttributeModifier("generic.attack_damage", tier_3_physical_multiplier/ 2, EntityAttributeModifier.Operation.MULTIPLY_BASE)
-            )
-    ),tier_3_spell_multiplier);
-    public static Entry vampire_ring = addLifeSteal(new Identifier(AdditionalJewelry.MOD_ID, "vampire_ring"), Rarity.RARE, true, new ItemConfig.Item(
-            List.of(
-            )
-    ),lifesteal_multiplier);
-    public static Entry vampire_necklace = addLifeSteal(new Identifier(AdditionalJewelry.MOD_ID, "vampire_necklace"), Rarity.RARE, true ,new ItemConfig.Item(
-            List.of(
-            )
-    ),lifesteal_multiplier);
-
+    private static final String WITCHER= "witcher_rpg";
     public static void register(ItemConfig allConfigs) {
+        if(FabricLoader.getInstance().isModLoaded(WITCHER)){
+            //RINGS
+            add(new Identifier(AdditionalJewelry.MOD_ID,"silver_sapphire_ring"),Rarity.UNCOMMON, new ItemConfig.Item(
+                    List.of(
+                            new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                                    Identifier.tryParse("witcher_rpg:adrenaline_modifier")),
+                                    tier_1_multiplier  , EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                            new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                                    Identifier.tryParse("witcher_rpg:sign_intensity")),
+                                    tier_1_multiplier, EntityAttributeModifier.Operation.MULTIPLY_BASE)
+
+                    )
+            ));
+            add(new Identifier(AdditionalJewelry.MOD_ID,"steel_jade_ring"),Rarity.UNCOMMON, new ItemConfig.Item(
+                    List.of(
+                            new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                                    Identifier.tryParse("witcher_rpg:adrenaline_modifier")),
+                                    0.08F , EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                            new ItemConfig.AttributeModifier("generic.attack_damage", tier_1_multiplier/2, EntityAttributeModifier.Operation.MULTIPLY_BASE)
+                    )
+            ));
+            add(new Identifier(AdditionalJewelry.MOD_ID, "dark_steel_jade_ring"), Rarity.UNCOMMON, new ItemConfig.Item(
+                    List.of(
+                            new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                                    Identifier.tryParse("witcher_rpg:adrenaline_modifier")),
+                                    0.14F , EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                            new ItemConfig.AttributeModifier("generic.attack_damage", tier_2_multiplier/2, EntityAttributeModifier.Operation.MULTIPLY_BASE)
+                    )
+            ));
+            add(new Identifier(AdditionalJewelry.MOD_ID, "meteorite_silver_sapphire_ring"), Rarity.UNCOMMON, new ItemConfig.Item(
+                    List.of(
+                            new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                                    Identifier.tryParse("witcher_rpg:adrenaline_modifier")),
+                                    0.1F , EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                            new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                                    Identifier.tryParse("witcher_rpg:sign_intensity")),
+                                    tier_2_multiplier, EntityAttributeModifier.Operation.MULTIPLY_BASE)
+                    )
+            ));
+            //NECKLACES
+            add(new Identifier(AdditionalJewelry.MOD_ID, "bear_school_medallion"), Rarity.RARE, true ,new ItemConfig.Item(
+                    List.of(
+                            new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                                    Identifier.tryParse("witcher_rpg:quen_intensity")),
+                                    2 , EntityAttributeModifier.Operation.ADDITION),
+                            new ItemConfig.AttributeModifier(GENERIC_MAX_HEALTH, 6.0F, EntityAttributeModifier.Operation.ADDITION),
+                            new ItemConfig.AttributeModifier("generic.attack_damage", tier_3_physical_multiplier, EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                            new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                                    Identifier.tryParse("witcher_rpg:adrenaline_modifier")),
+                                    0.2F , EntityAttributeModifier.Operation.MULTIPLY_BASE)
+                    )
+            ));
+            add(new Identifier(AdditionalJewelry.MOD_ID, "cat_school_medallion"), Rarity.RARE, true ,new ItemConfig.Item(
+                    List.of(
+                            new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                                    Identifier.tryParse("witcher_rpg:aard_intensity")),
+                                    2 , EntityAttributeModifier.Operation.ADDITION),
+                            new ItemConfig.AttributeModifier("generic.attack_speed", 0.06F, EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                            new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                                    Identifier.tryParse("witcher_rpg:adrenaline_modifier")),
+                                    0.2F , EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                            new ItemConfig.AttributeModifier("generic.attack_damage", tier_3_physical_multiplier,
+                                    EntityAttributeModifier.Operation.MULTIPLY_BASE)
+                    )
+            ));
+            add(new Identifier(AdditionalJewelry.MOD_ID, "griffin_school_medallion"), Rarity.UNCOMMON, new ItemConfig.Item(
+                    List.of(
+                            new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                                    Identifier.tryParse("witcher_rpg:sign_intensity")),
+                                    tier_3_spell_multiplier , EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                            new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                                    Identifier.tryParse("witcher_rpg:adrenaline_modifier")),
+                                    0.12F , EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                            new ItemConfig.AttributeModifier(SpellPowerMechanics.CRITICAL_DAMAGE.id, tier_3_secondary_multiplier, EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                            new ItemConfig.AttributeModifier(SpellPowerMechanics.HASTE.id, tier_3_secondary_multiplier, EntityAttributeModifier.Operation.MULTIPLY_BASE)
+                    )
+            ));
+            add(new Identifier(AdditionalJewelry.MOD_ID, "wolf_school_medallion"), Rarity.UNCOMMON, new ItemConfig.Item(
+                    List.of(
+                            new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                                    Identifier.tryParse("witcher_rpg:sign_intensity")),
+                                    tier_3_spell_multiplier , EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                            new ItemConfig.AttributeModifier(Objects.requireNonNull(
+                                    Identifier.tryParse("witcher_rpg:adrenaline_modifier")),
+                                    0.15F , EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                            new ItemConfig.AttributeModifier("generic.attack_speed", 0.06F, EntityAttributeModifier.Operation.MULTIPLY_BASE),
+                            new ItemConfig.AttributeModifier("generic.attack_damage", tier_3_physical_multiplier,
+                                    EntityAttributeModifier.Operation.MULTIPLY_BASE)
+                    )
+            ));
+
+        }
+
+
+
         for (var entry : all) {
             ItemConfig.Item itemConfig = allConfigs.items.get(entry.id.toString());
             if (itemConfig == null) {